Remove legacy code for conversation
