+++ front/components/assistant/conversation/constant.ts
@@ -1,3 +1,1 @@
-export const CONVERSATION_VIEW_SCROLL_LAYOUT = "conversation-scroll-layout";
-
 export const DEFAULT_RIGHT_PANEL_SIZE = 40;
